[TA-2675] Move parsing method to service
