+++ src/services/package-manager/variable-service.ts
@@ -6,8 +6,9 @@ import {
 } from "../../interfaces/package-manager.interfaces";
 import {variablesApi} from "../../api/variables-api";
 import { v4 as uuidv4 } from "uuid";
-import {logger} from "../../util/logger";
+import {FatalError, logger} from "../../util/logger";
 import {FileService, fileService} from "../file-service";
+import {URLSearchParams} from "url";
 
 class VariableService {
 
@@ -20,7 +21,8 @@ class VariableService {
     }
 
     public async listCandidateAssignments(type: string, params: string): Promise<void> {
-        const assignments = await variablesApi.getCandidateAssignments(type, params);
+        const parsedParams = this.parseParams(params);
+        const assignments = await variablesApi.getCandidateAssignments(type, parsedParams);
 
         assignments.forEach(assignment => {
             logger.info(JSON.stringify(assignment));
@@ -28,11 +30,29 @@ class VariableService {
     }
 
     public async findAndExportCandidateAssignments(type: string, params: string): Promise<void> {
-        const assignments = await variablesApi.getCandidateAssignments(type, params);
+        const parsedParams = this.parseParams(params);
+        const assignments = await variablesApi.getCandidateAssignments(type, parsedParams);
 
         const filename = uuidv4() + ".json";
         fileService.writeToFileWithGivenName(JSON.stringify(assignments), filename);
         logger.info(FileService.fileDownloadedMessage + filename);
+    }
+
+    private parseParams(params?: string): URLSearchParams {
+        const queryParams = new URLSearchParams();
+
+        if (params) {
+            try {
+                params.split(",").forEach((param: string) => {
+                    const paramKeyValuePair: string[] = param.split("=");
+                    queryParams.set(paramKeyValuePair[0], paramKeyValuePair[1]);
+                })
+            } catch (e) {
+                throw new FatalError(`Problem parsing query params: ${e}`);
+            }
+        }
+
+        return queryParams;
     }
 }
